% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpt_emmeans.R
\name{mpt_emmeans}
\alias{mpt_emmeans}
\title{emmeans()-wrapper for MPT models}
\usage{
mpt_emmeans(object, specs, by = NULL, type = "response", ..., simplify = TRUE)
}
\arguments{
\item{object}{fitted MPT model object returned by \code{\link[=mpt]{mpt()}}}

\item{specs}{same as for \code{\link[emmeans:emmeans]{emmeans::emmeans()}}}

\item{by}{same as for \code{\link[emmeans:emmeans]{emmeans::emmeans()}}}

\item{type}{same as for \code{\link[emmeans:summary.emmGrid]{emmeans::summary.emmGrid()}}. Default for MPT models
is \code{"response"}}

\item{...}{further arguments passed to \code{\link[emmeans:emmeans]{emmeans::emmeans()}} or corresponding
methods}

\item{simplify}{logical value indicating whether the different results should
be compiled into one \code{data.frame} (\code{TRUE}, the default) or returned as a
\code{list}}
}
\description{
A convenient interface to \code{\link[emmeans:emmeans]{emmeans::emmeans()}} for all MPT model parameters.
Per default, results for all MPT model parameters are combined into one
\code{data.frame}, this can be disabled using the \code{simplify} argument.
}
