% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp.R
\name{ppp}
\alias{ppp}
\title{Posterior Predictive P-values for MPTs}
\usage{
ppp(object, ndraws, type = "X2")
}
\arguments{
\item{object}{fitted MPT model object returned by \code{\link[=mpt]{mpt()}}.}

\item{ndraws}{number of posterior draws that should be used for calculating
the posterior predictive p-values.}

\item{type}{whether the T1 statistic of expected frequencies is computed using Person's "X2"(default) or the likelihood-ratio statistic "G2"}
}
\description{
Calculates posterior predictive p-values for MPT values. Currently only the
T1 statistic of Klauer (2010) is implemented.
}
\references{
Klauer, K. C. (2010). Hierarchical Multinomial Processing Tree
Models: A Latent-Trait Approach. \emph{Psychometrika}, 75(1), 70-98.
https://doi.org/10.1007/s11336-009-9141-0
}
