% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{skk13}
\alias{skk13}
\title{Singmann, Kellen, & Klauer (2013) ORE Data}
\format{
\subsection{skk13}{

A data frame with 8,400 rows and 7 variables:
\describe{
\item{id}{Factor with 42 levels indicating the participant identifier.}
\item{trial}{Factor indicating the trial number.}
\item{race}{Factor indicating the race of the image shown, either
\code{german} (own-race) or \code{arabic} (other-race).}
\item{type}{Factor indicating the item type (i.e., MPT tree), either \code{old}
(i.e., studied item) or \code{new} (i.e., non-studied item).}
\item{resp}{Factor indicating the response, either \code{old}, \code{unsure}, or \code{new}.}
\item{rt}{response time in seconds.}
\item{stim}{Factor with 200 levels indicating the shown image/item.}
}
}
}
\source{
Singmann, H., Kellen, D., & Klauer, K. C. (2013). Investigating the Other-Race Effect of Germans towards Turks and Arabs using Multinomial Processing Tree Models. In M. Knauff, M. Pauen, N. Sebanz, & I. Wachsmuth (Eds.), Proceedings of the 35th Annual Conference of the Cognitive Science Society (pp. 1330–1335). Austin, TX: Cognitive Science Society.
\url{http://singmann.org/download/publications/SKK-CogSci2013.pdf}
}
\usage{
skk13
}
\description{
Data presented in Singmann, Kellen, & Klauer (2013) investigating the other
race effect (ORE) using the unsure-extended 2-high threshold model.
Participants perform a face recognition memory test for both own-race and
other-race faces. The original analysis is performed using no-pooling maximum
likelihood estimation and shows an effect of target race on the \code{Dn}
parameter.
}
\details{
The original abstract is as follows: The other-race effect (ORE) refers to the phenomenon that recognition memory for other-race faces is worse than for own-race faces. We investigated whether White Germans exhibited an ORE towards Turkish or Arabic faces using a multinomial processing tree model (MPT), the two-high threshold model of recognition memory with three response categories (old, skip, and new). Using an MPT enabled us to adequately disentangle memory and response processes using the Fisher information approximation, a minimum description length based measure of model complexity. Results showed that participants exhibited an ORE on the memory parameters but not on the parameters representing response processes.
}
\keyword{datasets}
